sam-indexing/requirements.txt
boto3==1.20.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-api-python-client==2.30.0
requests==2.26.0
